🚚 Rename: mapper -> converter로 구조 통일
