+++ src/main/java/project/masil/auth/converter/AuthConverter.java
@@ -1,11 +1,11 @@
-package project.masil.auth.mapper;
+package project.masil.auth.converter;
 
 import org.springframework.stereotype.Component;
 import project.masil.auth.dto.response.LoginResponse;
 import project.masil.user.entity.User;
 
 @Component
-public class AuthMapper {
+public class AuthConverter {
 
   public LoginResponse toLoginResponse(User user, String accessToken, Long expirationTime) {
     return LoginResponse.builder()
